# feat: LinkedIn MCP Server v1.6.0 - Performance Revolution with Centralized Stealth Architecture