--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "linkedin-mcp-server"
-version = "1.5.0"
+version = "1.6.0"
 description = "Optimized LinkedIn MCP Server"
 requires-python = ">=3.12"
 dependencies = [