… README.md
… RELEASE_NOTES.md
… RELEASE_NOTES_v1.5.0.md
… linkedin_mcp_server/scraper/config.py
… linkedin_mcp_server/scraper/pages/base.py
… linkedin_mcp_server/scraper/pages/profile_page.py
… linkedin_mcp_server/scraper/scrapers/person/accomplishments.py
… linkedin_mcp_server/scraper/scrapers/person/connections.py
… linkedin_mcp_server/scraper/scrapers/person/contacts.py
… linkedin_mcp_server/scraper/scrapers/person/education.py
… linkedin_mcp_server/scraper/scrapers/person/experience.py
… linkedin_mcp_server/scraper/scrapers/person/get_person.py
… linkedin_mcp_server/scraper/scrapers/person/interests.py
… linkedin_mcp_server/scraper/scrapers/person/selectors.py
… linkedin_mcp_server/scraper/scrapers/person/utils.py
… linkedin_mcp_server/scraper/stealth/__init__.py
… linkedin_mcp_server/scraper/stealth/controller.py
… linkedin_mcp_server/scraper/stealth/hooks.py
… linkedin_mcp_server/scraper/stealth/lazy_loading.py
… linkedin_mcp_server/scraper/stealth/navigation.py
… linkedin_mcp_server/scraper/stealth/profiles.py
… linkedin_mcp_server/scraper/stealth/simulation.py
… linkedin_mcp_server/scraper/stealth/telemetry.py
… linkedin_mcp_server/tools/person.py
… tests/integration/test_stealth_performance.py
… tests/unit/stealth/test_stealth_profiles.py
… uv.lock